Update 12.0u1
New features:
- Added full UEFI support
- Automatic sshkeypair generation for users without an existing keypair

Changes:
- Moved configuration files to .example files, in order to have a care-free update
- Appended configurationfiles and buildables to gitignore
- Updated remote assistance files to python3
- Added remote assistance systemd service
- Modified Trinity Health checks and renamed to Trinity Diag (trix-diag)
- Updated README.rst
- Updated prepare.sh with dependencies

Fixes:
- Fixed documentation layout
- Fixed InfluxDB retention
- Fixed non-existent python3 env on clean installations
- Fixed compute image generation (compute.yml)
- Fixed rsyslog forwarding

+++ prepare.sh
@@ -1,10 +1,13 @@
 #!/bin/bash
-setenforce 0
-yum update -y
-curl https://updates.clustervision.com/luna/1.2/centos/luna-1.2.repo > /etc/yum.repos.d/luna-1.2.repo
-yum install epel-release -y
-yum install git ansible luna-ansible -y
-ssh-keygen -q -t rsa -b 4096
-cat ~/.ssh/id_rsa.pub >> ~/.ssh/authorized_keys
-ansible-galaxy install OndrejHome.pcs-modules-2
-echo "#### Please configure the network before starting Ansible ####"
+if [ ! -f TrinityX.pdf ]; then
+  echo "Please run from within the cloned TrinityX folder"
+else
+  setenforce 0
+  yum update -y
+  curl https://updates.clustervision.com/luna/1.2/centos/luna-1.2.repo > /etc/yum.repos.d/luna-1.2.repo
+  yum install epel-release -y
+  yum install ansible git luna-ansible python-pip -y
+  ansible-galaxy install OndrejHome.pcs-modules-2
+  git clone https://github.com/dw/mitogen.git site/mitogen
+  echo "#### Please configure the network before starting Ansible ####"
+fi
